✅ Passing 09 and 10 tests. Fixing RLCA and SWAP. Moved DEC in correct file.
