+++ tests/README.md
@@ -9,9 +9,9 @@
 - [x] 06-ld r,r.gb
 - [x] 07-jr,jp,call,ret,rst.gb
 - [x] 08-misc instrs.gb
-- [ ] 09-op r,r.gb
+- [x] 09-op r,r.gb
 - [x] 10-bit ops.gb
-- [ ] 11-op a,(hl).gb
+- [x] 11-op a,(hl).gb
 
 ## instr_timing
